Rename Spring AOP integration test modules for clarity
