+++ internal/micrometer-tracing-integration-test-spring-aop-no-aspectj/build.gradle
@@ -5,7 +5,6 @@ plugins {
 dependencies {
 	testImplementation project(':micrometer-tracing')
 	testImplementation project(':micrometer-tracing-test')
-	testImplementation libs.springContext
-	testImplementation libs.aspectjweaver
+	testImplementation libs.springAop
 	testImplementation libs.assertj
 }
