+++ settings.gradle
@@ -39,7 +39,7 @@ include 'micrometer-tracing', 'micrometer-tracing-bom', 'benchmarks'
 	project(":micrometer-tracing-$testProject").projectDir = new File(rootProject.projectDir, "micrometer-tracing-tests/micrometer-tracing-$testProject")
 }
 
-['integration-test-spring-aop', 'integration-test-jpms', 'integration-test-no-aspectj'].each { testProject ->
+['integration-test-spring-aop-with-aspectj', 'integration-test-spring-aop-no-aspectj', 'integration-test-jpms'].each { testProject ->
 	include "micrometer-tracing-$testProject"
 	project(":micrometer-tracing-$testProject").projectDir = new File(rootProject.projectDir, "internal/micrometer-tracing-$testProject")
 }
